+++ linter/lints/crl/lints.go
@@ -114,23 +114,6 @@ func noZeroReasonCodes(crl *crl_x509.RevocationList) *lint.LintResult {
 	return &lint.LintResult{Status: lint.Pass}
 }
 
-// noCrticialReasons checks Baseline Requirements, Section 7.2.2.1:
-// If present, [the reasonCode] extension MUST NOT be marked critical.
-func noCriticalReasons(crl *crl_x509.RevocationList) *lint.LintResult {
-	reasonCodeOID := asn1.ObjectIdentifier{2, 5, 29, 21} // id-ce-reasonCode
-	for _, rc := range crl.RevokedCertificates {
-		for _, ext := range rc.Extensions {
-			if ext.Id.Equal(reasonCodeOID) && ext.Critical {
-				return &lint.LintResult{
-					Status:  lint.Error,
-					Details: "CRL entry reasonCodes MUST NOT be critical",
-				}
-			}
-		}
-	}
-	return &lint.LintResult{Status: lint.Pass}
-}
-
 // noCertificateHolds checks Baseline Requirements, Section 7.2.2.1:
 // The CRLReason MUST NOT be certificateHold (6).
 func noCertificateHolds(crl *crl_x509.RevocationList) *lint.LintResult {
